정수 대신 Short 사용 코드경감

+++ src/main/java/tacos/Taco.java
@@ -26,7 +26,7 @@ public class Taco {
   @NotNull
   @Size(min = 1, message = "재료를 한 가지 이상 선택하세요.")
   @Transient
-  private List<Integer> ingredientSNs = new ArrayList<>();
+  private List<Short> ingredientSNs = new ArrayList<>();
 
   private List<IngredientRef> ingredientRefs = new ArrayList<>();
